requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    inv pin
#
attrs==19.1.0             # via jsonschema, packaging
certifi==2019.6.16        # via requests
chardet==3.0.4            # via requests
coreapi==2.3.3            # via drf-yasg
coreschema==0.0.4         # via coreapi, drf-yasg
django-filter==2.2.0
django==2.2
djangorestframework==3.9.0
drf-yasg[validation]==1.16.1
idna==2.8                 # via requests
inflection==0.3.1         # via drf-yasg
invoke==1.3.0
itypes==1.1.0             # via coreapi
jinja2==2.10.1            # via coreschema
jsonschema==3.0.2         # via swagger-spec-validator
markdown==3.1.1
markupsafe==1.1.1         # via jinja2
packaging==19.1
pyparsing==2.4.2          # via packaging
pyrsistent==0.15.4        # via jsonschema
pytz==2019.2              # via django
pyyaml==5.1.2             # via swagger-spec-validator
requests==2.22.0          # via coreapi
ruamel.yaml.clib==0.1.2   # via ruamel.yaml
ruamel.yaml==0.16.5       # via drf-yasg
six==1.12.0               # via drf-yasg, jsonschema, packaging, pyrsistent, swagger-spec-validator
sqlparse==0.3.0           # via django
swagger-spec-validator==2.4.3  # via drf-yasg
uritemplate==3.0.0        # via coreapi, drf-yasg
urllib3==1.25.3           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.2.0        # via jsonschema, markdown
